ocr/requirements.txt
fastapi
uvicorn[standard]
python-multipart
sqlalchemy
psycopg2-binary
python-dotenv
easyocr
reportlab
Pillow
numpy
torch
torchvision
transformers
opencv-python
timm
requests
